+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Changes
   * (Closes #108)
 * Consistently throw exceptions in HTN generation using fatal-error
 * Ensure metadata in TPN's flows from method definitions when generated with the HTN method (Closes #98)
+* Updated plan-schema to leverage the match-eval-out-err macro in testing
 * Ensure remove-superfluous in TPN optimization does not re-visit nodes
   unnecessarily (Closes #115)